requirements.txt
Flask==2.3.3
schedule==1.2.0
python-dotenv==1.0.0